Move cmdline arg parsing to a utility function
