+++ stix2validator/__init__.py
@@ -6,7 +6,7 @@ import sys
 
 from .errors import ValidationError
 from .output import print_results
-from .util import ValidationOptions
+from .util import ValidationOptions, parse_args
 from .validator import (run_validation, validate, validate_file,
                         validate_parsed_json, validate_string)
 from .version import __version__
